Move CauseDetail containers
