+++ src/components/containers/CauseDetail/AboutCause/AboutCause.spec.js
