+++ src/components/containers/CauseDetail/ProgressBar/ProgressBar.js
