+++ src/components/containers/CauseDetail/ProgressBar/ProgressBar.spec.js
